requirements.txt
flake8==3.7.9
yapf==0.29.0
isort==4.3.21
yacs
gdown
tb-nightly
future
scipy
scikit-learn
tqdm
ftfy
regex
wilds==1.2.2
tabulate
torch 
torchvision
transformers
SentencePiece
PyYAML
Pillow
numpy
open-clip-torch